requirements.txt
fastapi==0.118.0
uvicorn[standard]==0.37.0
jinja2==3.1.6
python-multipart==0.0.20
pillow==12.0.0
pandas==2.3.2
folium==0.20.0
requests==2.32.5
openpyxl==3.1.2
